Add concentration parameters of air pollutants.
At the moment there is not available any concentration parameters
of air pollutants in the library. These additions makes possible to
store mass per unit volume forecast data of air pollutants.

Naming of a parameter does not take account which is the unit of data
but it must be an unit of concentration. Unit can be for example
milligrams of pollutant per cubic meter.