Unset GDAL from Cmake Cache if it is disabled afterwards.
If a user run cmake without giving option -DDISABLE_GDAL=ON, GDAL
setting will be cached to the CMakeCache.txt file because it is enabled
by default. If the user want to disable it later, the settings stays
in the cache and are used on compilation. The shared version of newbase
is then depended on GDAL library.