Added parameters GrassFireIndex and ForestFireIndex (SOL-7213). Changed some NFmiFastQueryInfo metods to be public.
Iterating data via a single NFmiFastQueryInfo instance is not thread
safe using the public API. Using the indexes directly however is
safe. This enables us to for example run the road model in parallel
for multiple coordinates *and* to store the results into querydata
in parallel without a new NFmiFastQueryInfo instance for each thread.