Fixed ToLatLon to try to return the desired longitude if the result is very close to a pole
GDAL/PROJ.4 will set the longitude to zero if the result is very close
to a pole. If we detect this to be the case within an epsilon
(0.0001), we try to project the original coordinate again but with the
metric Y-coordinate set to zero. The output longitude will be taken
from this auxiliary result.

For example, it is common for global forecasts to have
-180,-90...180,90 as the bounding box. If the latlon bounding box is
calculated from the metric bounding box (proj=eqc), the result may be
0,-90...0,90. For proj=eqc the X coordinate is the same for all
longitudes, hence performing the same calculation with Y=0 instead
will return the longitude we desire. The situation is not so simple
for other projections, but setting any longitude for poles is still
essentially the same coordinate.