Removed NFmiFastQueryInfo::Locations which returns latlon coordinates as dangerous
Removing the templated method will reveal code which potentially will
reproject the geographic coordinates further. The correct way is to
request the CoordinateMatrix and project it directly all the way to
the target projection without going through WGS84 coordinates.

It would be possible to add a Locations method which would return a
CoordinateMatrix in WGS84 coordinates later on, but it is better to do
so once all locations revealed by this removal have been checked for
correctness.

The code would then be:

  NFmiCoordinateTransformation transformation(SpatialReference(), "WGS84");
  auto coords = CoordinateMatrix();
  coords.Transform(transformation);
  return coords;